refactor: Reorganize tests directory
