+++ TODO.md
@@ -1,1 +1,8 @@
+- branch for custom matchers
+- pretty
+- multiline
+- whitespaces
+- drop JSON
 - get rid of "extend" dependency
+- handle "not"
+
