+++ demo/matcher-to-equal.spec.js
@@ -49,27 +49,18 @@ describe('Matcher', function () {
       expect(a).toEqual(jasmine.objectContaining({ foo: 43 }));
     });
 
-    fit('should NOT diff with jasmine.objectContaining', function () {
-      // var a = {
-      //   foo: function () {},
-      //   bar: function () {}
-      // };
-      // var b = {
-      //   foo: function () {},
-      //   bar: jasmine.any(Function)
-      // };
-      // var a = function(){};
-      // var b = function() {};
-      // expect(a).toEqual(b);
-
-      function Foo() {
-        this.bar = 42;
-        this.hello = function () {};
-      }
-
-      var a = 3;
-      var b = 3;
-      expect(new Foo()).toEqual(jasmine.any(Object));
+    it('should NOT diff with jasmine.objectContaining', function () {
+      var a = {
+        foo: function () {},
+        bar: function () {}
+      };
+      var b = {
+        foo: function () {},
+        bar: jasmine.any(Function)
+      };
+      var a = function(){};
+      var b = function() {};
+      expect(a).toEqual(b);
     });
   });
