+++ test/exclusions.test.js
@@ -1,0 +1,10 @@
+var createTest = require('./helpers/test');
+
+var test = createTest('exclusions:');
+
+test('unmatched message',
+
+  `Some random text`,
+
+  `Some random text`
+);
