--- test/helpers/stack-wrap.js
@@ -1,7 +1,0 @@
-'use strict';
-
-var stack = require('../helpers/stack');
-
-module.exports = function (str) {
-  return str + stack;
-};
