+++ test/helpers/test.js
@@ -4,25 +4,30 @@ var tape = require('tape');
 
 var format = require('../../src/format');
 var formatter = require('./test-formatter')();
-var s = require('./stack-wrap');
-var m = require('../../src/marker').wrapString;
+var stack = require('./stack');
 
 module.exports = function createTestFn(namespace) {
 
-  var test = function (name, input, expected) {
-    return tape(namespace + ' ' + name,  function (assert) {
-      var out = format(s(input), formatter);
+  var test = function (name, input, expected, options) {
+    options = options || {};
+    options.stack = options.stack || stack;
 
-      assert.equal(out, s(expected));
+    return tape(namespace + ' ' + name,  function (assert) {
+      var out = format(input + stack, formatter);
+
+      assert.equal(out, expected + stack);
       assert.end();
     });
   };
 
-  test.only = function (name, input, expected) {
-    return tape.only(namespace + ' ' + name,  function (assert) {
-      var out = format(s(input), formatter);
+  test.only = function (name, input, expected, options) {
+    options = options || {};
+    options.stack = options.stack || stack;
 
-      assert.equal(out, s(expected));
+    return tape.only(namespace + ' ' + name,  function (assert) {
+      var out = format(input + stack, formatter);
+
+      assert.equal(out, expected + stack);
       assert.end();
     });
   };
